# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LARAVEL Rest API DOCUMENTATION",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "manirujjamanakash@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Auth Login",
                "description": "Auth Login",
                "operationId": "eaf28e54bd4a4126b9668fe26e6d1411",
                "requestBody": {
                    "description": "User objects",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "email",
                                        "type": "string",
                                        "example": "gugu@gmail.com"
                                    },
                                    "password": {
                                        "description": "password",
                                        "type": "string",
                                        "example": "sithuadmin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "description": "Register",
                "operationId": "a3f252b10cbf8e270739c3ada00c6042",
                "requestBody": {
                    "description": "User objects",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "name",
                                        "type": "string",
                                        "example": "gugu"
                                    },
                                    "email": {
                                        "description": "email",
                                        "type": "string",
                                        "example": "gugu@gmail.com"
                                    },
                                    "password": {
                                        "description": "password",
                                        "type": "string",
                                        "example": "guguadmin"
                                    },
                                    "password_confirmation": {
                                        "description": "password_confirmation",
                                        "type": "string",
                                        "example": "guguadmin"
                                    },
                                    "image": {
                                        "description": "image",
                                        "type": "file",
                                        "example": "image"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/comments": {
            "post": {
                "tags": [
                    "Comment"
                ],
                "summary": "Comment Create",
                "description": "Create Comment",
                "operationId": "917b06c80f173dc9400c056742570b45",
                "requestBody": {
                    "description": "Comment objects",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "movie_id": {
                                        "description": "movie_id",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "comment": {
                                        "description": "comment",
                                        "type": "string",
                                        "example": "good"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/comments/list": {
            "get": {
                "tags": [
                    "Comment"
                ],
                "summary": "Get Comment",
                "description": "Get Comment",
                "operationId": "bea1468e48f7ab9b49c2f475735c9d6e",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "description": "Enter your bearer token in the format `Bearer {token}`"
            }
        }
    }
}